add pagination

+++ app/http/controllers/v1/role/role_controller.go
@@ -2,8 +2,8 @@ package role
 
 import (
 	"github.com/goravel/framework/contracts/http"
+	"karuhundeveloper.com/gostarterkit/app/helpers"
 	request "karuhundeveloper.com/gostarterkit/app/http/requests/v1/role"
-	"karuhundeveloper.com/gostarterkit/app/http/responses"
 	"karuhundeveloper.com/gostarterkit/app/services/v1/role"
 )
 
@@ -18,32 +18,45 @@ func NewRoleController(RoleService *role.RoleService) *RoleController {
 	}
 }
 
+func (r *RoleController) Index(ctx http.Context) http.Response {
+	if !helpers.HasPermission(ctx, "view_all_role") {
+		return helpers.ErrorResponse(ctx, http.StatusForbidden, "Validation Error", helpers.NoPermissionError().Error())
+	}
+
+	// List roles
+	roles, pagination, err := r.roleService.List(ctx)
+
+	if err != nil {
+		return helpers.ErrorResponse(ctx, http.StatusBadRequest, "Role Listing Failed", err.Error())
+	}
+
+	return helpers.SuccessResponse(ctx, http.StatusOK, "Roles Retrieved Successfully", http.Json{
+		"data":       roles,
+		"pagination": pagination,
+	});
+}
+
 func (r *RoleController) Create(ctx http.Context) http.Response {
 	var roleCreateRequest request.RoleCreateRequest
 
 	// Validate request data
 	validationErrors, err := ctx.Request().ValidateRequest(&roleCreateRequest)
 	if err != nil {
-		return responses.ErrorResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", err.Error())
+		return helpers.ErrorResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", err.Error())
 	}
 
 	if validationErrors != nil {
-		return responses.ErrorValidationResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", validationErrors.All())
+		return helpers.ErrorValidationResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", validationErrors.All())
 	}
 
 	// Create role
 	role, err := r.roleService.Create(ctx, roleCreateRequest)
 
 	if err != nil {
-		return responses.ErrorResponse(ctx, http.StatusInternalServerError, "Role Creation Failed", err.Error())
+		return helpers.ErrorResponse(ctx, http.StatusBadRequest, "Role Creation Failed", err.Error())
 	}
 
-	return responses.SuccessResponse(ctx, http.StatusCreated, "Role Created Successfully", http.Json{
-		"id":   role.ID,
-		"name": role.Name,
-		"created_at": role.CreatedAt,
-		"updated_at": role.UpdatedAt,
-	})
+	return helpers.SuccessResponse(ctx, http.StatusCreated, "Role Created Successfully", helpers.ModelToMap(role))
 }
 
 func (r *RoleController) Update(ctx http.Context) http.Response {
@@ -52,35 +65,35 @@ func (r *RoleController) Update(ctx http.Context) http.Response {
 	// Validate request data
 	validationErrors, err := ctx.Request().ValidateRequest(&roleUpdateRequest)
 	if err != nil {
-		return responses.ErrorResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", err.Error())
+		return helpers.ErrorResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", err.Error())
 	}
 
 	if validationErrors != nil {
-		return responses.ErrorValidationResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", validationErrors.All())
+		return helpers.ErrorValidationResponse(ctx, http.StatusUnprocessableEntity, "Validation Error", validationErrors.All())
 	}
 
 	// Update role
 	role, err := r.roleService.Update(ctx, roleUpdateRequest)
 
 	if err != nil {
-		return responses.ErrorResponse(ctx, http.StatusInternalServerError, "Role Update Failed", err.Error())
+		return helpers.ErrorResponse(ctx, http.StatusNotFound, "Role not found", err.Error())
 	}
 
-	return responses.SuccessResponse(ctx, http.StatusOK, "Role Updated Successfully", http.Json{
-		"id":   role.ID,
-		"name": role.Name,
-		"created_at": role.CreatedAt,
-		"updated_at": role.UpdatedAt,
-	})
+	return helpers.SuccessResponse(ctx, http.StatusOK, "Role Updated Successfully", helpers.ModelToMap(role))
 }
 
 func (r *RoleController) Delete(ctx http.Context) http.Response {
+	// Check permission
+	if (!helpers.HasPermission(ctx, "delete_role")) {
+		return helpers.ErrorResponse(ctx, http.StatusForbidden, "Validation Error", helpers.NoPermissionError().Error())
+	}
+
 	// Delete role
 	err := r.roleService.Delete(ctx)
 
 	if err != nil {
-		return responses.ErrorResponse(ctx, http.StatusInternalServerError, "Role Deletion Failed", err.Error())
+		return helpers.ErrorResponse(ctx, http.StatusNotFound, "Role Deletion Failed", err.Error())
 	}
 
-	return responses.SuccessResponse(ctx, http.StatusOK, "Role Deleted Successfully", nil)
+	return helpers.SuccessResponse(ctx, http.StatusOK, "Role Deleted Successfully", nil)
 }
